Update patch set 8
Patch Set 8: Code-Review-1

(1 comment)

Looks nice, one more suggestion, we might as well use "public final" for the 4 methods in MessageNano as it looks like I was wrong. I found that dexdump -f prints out the dex header which as a method_ids_size and file_size fields so I tested various configurations and generated the text version of the dex file using the following:

  $ dx --dex --keep-classes --output=NanoTest.dex com/google/protobuf/NanoTest.class
  $ dexdump -f d NanoTest.dex > NanoText.txt


I then looked at "public static final" as in patch set 8 and we see the following:

  17:40:57 wink@wsaville /mnt/disk4/i4jb-mr2-dev/external/protobuf/java/target-ps8/test-classes(nano-ps8)
  $ f f '*Nano*txt' | xargs grep 'method_ids_size\|file_size'
  ./SimpleMessageNano.txt:file_size           : 2876
  ./SimpleMessageNano.txt:method_ids_size     : 21
  ./NanoTest.txt:file_size           : 33036
  ./NanoTest.txt:method_ids_size     : 100
  ./NanoOuterClass-TestAllTypesNano.txt:file_size           : 22136
  ./NanoOuterClass-TestAllTypesNano.txt:method_ids_size     : 100


When just being "public final" we got two extra methods in NanoTest although the file sizes are ever so slightly smaller in all three cases:

  $ f f '*Nano*txt' | xargs grep 'method_ids_size\|file_size'
  ./SimpleMessageNano.txt:file_size           : 2864
  ./SimpleMessageNano.txt:method_ids_size     : 21
  ./NanoTest.txt:file_size           : 32996
  ./NanoTest.txt:method_ids_size     : 102
  ./NanoOuterClass-TestAllTypesNano.txt:file_size           : 22124
  ./NanoOuterClass-TestAllTypesNano.txt:method_ids_size     : 100


And as another data point the micro version has up to 4x the number of methods and always smaller files. On the Super Jumbo TestAllType class, nano is half the file size.

  $ f f '*Micro*txt' | xargs grep 'method_ids_size\|file_size'
  ./SimpleMessageMicro.txt:file_size           : 3852
  ./SimpleMessageMicro.txt:method_ids_size     : 34
  ./MicroTest.txt:file_size           : 41636
  ./MicroTest.txt:method_ids_size     : 390
  ./MicroOuterClass-TestAllTypesMicro.txt:file_size           : 50156
  ./MicroOuterClass-TestAllTypesMicro.txt:method_ids_size     : 426


So I suggest we change the 4 "public stack final" methods in MessageNano to "public final" as you'd suggested. Ulas, I can push up a change if you'd like or you can do it.

Patch-set: 8
Label: Code-Review=-1