Update patch set 6
Patch Set 6:

PS4 = PS2 + remove an unnecessary TODO + mirror to nano. PS5-PS6 contain only typo/documentation fixes. PS6 ready to review. Making a clean full internal master build with PS6 for verification.

The updated behavior info (in FileGenerator::Validate) does not break the compilation, because 1) java_out allows this case, 2) it's quite frequent that the file class name will clash with the single message's name, so there will be a compile error, and 3) even if it succeeds, the message class is now in an outer class, so the main project will break.

Patch-set: 6
Label: Verified=0