Update patch set 2
Patch Set 2:

This was motivated by reducing memory usage for a common use case that is unique to nano.

Clients would like to pre-allocate an array of messages and fill it in over time, then set this into a repeated message field. We would like to avoid having to copy the array to the right size. Instead, just checking for null elements (in this use case really only at the tail of the array) is better.

Patch-set: 2