Update patch set 2
Patch Set 2:

APIs that are not nullable and enforce the lack of nullability through precondition checks are good (because they fail fast). APIs that fail slow like this are more problematic.

We fixed this downstream because we had actual cases in our app which triggered these NPEs. Trying to audit all our code to make sure nobody is setting repeated fields to null, and later serialization or toString is not causing an NPE is hard, as is enforcing that nobody breaks this again in future.

I think it's probably more problematic with the reftypes mode, because people are accustomed to null being equivalent to a clear / unset field in that mode. I can make this reftype specific if you like, but it seems like if we're checking in that mode, we should just always check.

Patch-set: 2
Label: Verified=0