Update patch set 1
Patch Set 1:

fwiw, we have a generator option in the downstream version that strips the clear methods, because we never use them, and didn't want to pay for their unproguarded cost toward the method count limit.

Not stripping clear() would increase our pre-proguard method count by ~2,500, which is a bit sad for us.

That said, I can always take a look later at preserving the option and just making the generator inline the clear code in the constructor when the clear method is not generated.

Patch-set: 1