Update patch set 1
Patch Set 1:

Local 'result' is fine; it's just the field called 'result' should be accessed through 'this.result' to disambig. Also, depending on various circumstances, the local variable could be called 'result_' (and in this case you don't need the 'this' qualifier). All in all, that really means the *only* way to go is to delegate the codegen to the individual FieldGenerators (FieldGenerator::GenerateEquals / GenerateHashCode), which will know what exactly to do.

Patch-set: 1