Update patch set 4
Patch Set 4:

Sounds like there are no concerns with any of the code changes here, but that there may be more questions about switching to toText().

Based on some brief reading about proguard, it does not detect reflection and disable method removal -- attempts to use method reflection without the appropriate configuration will fail.

Assuming this is true, the overhead of keeping toString() implemented via MessageNanoPrinter is that MessageNanoPrinter will always be retained. This would be fairly minimal (a class, a handful of methods, and the dependencies of it).

I propose we back out the deprecation of toString and the addition of toText until we have data on how proguard handles this. Max, it sounded like you were going to be doing some experimentation with proguard soon, is this something you could look into at the same time? I've uploaded another PS that does this, plus it if you're OK with this.

Patch-set: 4
Label: Verified=0