Update patch set 2
Patch Set 2:

Why are we dropping the enum if it doesn't match one of the known cases?  If it's an unknown value, it should default to the first value in the enum.  This is why we recommend adding an UNKNOWN value as the first enum (go/protosdosdonts #5).  I think we need this to prevent new enum values from breaking older clients in the wild.

Technically in the defaulting scenario, hasEnum() should return false as well, but I'm not sure how people are using hazzers with nano.

Patch-set: 2
Reviewer: Thomas Chao <1022933@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0