Update patch set 7
Patch Set 7: Verified-1

In trying this out I realized that build.gradle needs to be updated as well.

Unfortunately, this doesn't look completely trivial since the current build.gradle builds a plain jar, and introducing the Android dependency is a bit trickier than doing so with Android.mk. Will have to investigate this.

Patch-set: 7
Label: Verified=-1