Update patch set 11
Patch Set 11: Verified+1

OK - I've taken a stab at updating build.gradle to handle the new Android dependencies and verified that it works in an external Android project with this dependency:

compile project(path: ":protolib", configuration: 'nano')

xav@, would you mind taking a look and seeing if what I've done is reasonable? In a nutshell, rather than building a vanilla .jar, I now build an Android library project and copy the output class files into a .jar. The resulting output seems to be the same as before, but now the .class files which depend on the Android SDK are there.

This would break anyone depending on our build.gradle that isn't using the Android SDK; I doubt that this is anyone. However, if there's concern, I could instead leave nano as is and make a new nanoAndroid jar with just the Android components.

Patch-set: 11
Label: Verified=+1