Keep pointers to extension values.
The current implementation of getExtension deserialises the field from bytes
and returns a new object every time. This means that changes to those objects
are reflected when the messages is serialised unless setExtension is called. It
also means that every call to getExtension and setExtension is expensive.

This change adds fields in UnknownExtensionData class that can be used to keep a
pointer to the Extension and value when getting and setting extensions.

This means that:
* calling getExtension more than once deserialises the field only once and
  returns the same object.
* calling setExtension doesn't cause the object to be serialised immediately,
  that only happens when the overall message is serialised.

Bug: 10863158

Change-Id: I69b564410dc7297567157fa4d1b13fd5dbe5d67c

Signed-off-by: Juan Silveira <jjso@google.com>