Update patch set 2
Patch Set 2: Code-Review-1

So there was no need to change any of the nano or micro code generators to support the new features? And that the existing code generators work without any changes!

In any case its a good start.

One other issue is that there is internal code that is expecting to link against 2.3 libraries. The build system, in many cases, can supply the necessary library so we could change existing code to not expect a particular library first and then less stuff breaks when this goes in. Alternatively, we can just change the explicit references to 2.6.

Patch-set: 2
Label: Code-Review=-1