Update patch set 3
Patch Set 3:

Thanks - I personally only want this for unbundled apps so I'm not as concerned about the version in L branches. It could always be cherry-picked back if someone has a need for it.

Let me patch into master and see what we get. It'd be hard for me to check unbundled branches - I think the best I can do there is send out a warning noting that unbundled manifests should be pinned to a specific commit, along the lines of Max's previous "PSA: Protobuf Nano: big changes ahead; pin to ToT if needed" post.

Patch-set: 3