Update patch set 3
Patch Set 3:

I'd also honestly be fine just leaving the 2.3.0 in there. It's unclear to me why it was put there given that no other library in the Android tree seems to refer to itself by version code, and there are a *lot* of things that would need to be bumped each time given that we don't really have a mechanism for multiple versions of anything in the tree.

Patch-set: 3
Label: Verified=0