Update patch set 3
Patch Set 3:

> So there are a fair number of places internally which depend on
 > libprotobuf 2.3.0 by name, but likely not by actual functionality.
 > 
 > What would you think about removing the version code from the
 > outputted library completely? Then we could do future bumps without
 > mucking with every makefile that depends on protobufs.

For those places, do they use the library directly or just include it because they're compiling proto files? If it's always the latter, I think it might be helpful to make the 2.3.0 target an empty target, to limit the scope of the changes. The 2.6.0 version is included by the build system anyway so nothing should break.

Patch-set: 3