Update patch set 3
Patch Set 3: Code-Review+1

Is the plan to provide some instructions, probably in external/protobuf/java/README.txt, on the best practices on how to use protobufs in android, how to "properly" upgrade from 2.3 to 2.6 and how to use 2.3 if necessary?

Should we mention how to use LOCAL_OVERRIDE_PACKAGES here?

Patch-set: 3