Update patch set 3
Patch Set 3: Code-Review+1 Verified+1

I did a successful full-eng build with https://android-review.googlesource.com/#/c/107725 as well. This means that LOCAL_OVERRIDES_PACKAGES is working correctly (stripping out the 2.3.0 lib if someone is specifying both it and the proto type, which would pull in the newer lib), and that the new compiler isn't breaking anyone in AOSP.

@Wink: In the change which actually updates the compiler, I will also include instructions on the current state of the world. (Note that the prebuilts/Android.mk does mention that these are deprecated, but I'll make a broader note in java/README.txt).

Re: how to use LOCAL_OVERRIDES_PACKAGES: I'm not sure what you mean. Nobody depending on protobufs should need to know how to use this (for protobufs, at least). All this does is ensure that if the make rule is depending on both libprotobuf-*-2.3.0-* and libprotobuf-*-*, then the versionless one takes precedence and overrides the dependency on the old prebuilt.

Patch-set: 3