Update patch set 4
Patch Set 4:

(3 comments)

I have also updated https://android-review.googlesource.com/#/c/107725/ as discussed. If a makefile contains both a LOCAL_PROTOC_OPTIMIZE_TYPE and an explicit dependency on one of the 2.3.0 protolibs, we log a warning and strip the 2.3.0 dependency.

Note: my plan is to submit this first, wait a day or so to ensure no breakage, send a note out to android-eng notifying about upcoming changes and how to address them, and then submit the other changes the next day or so.

Patch-set: 4
Label: Verified=0