Update patch set 5
Patch Set 5:

Well, unfortunately it looks like I hit a snag. The C++ apis are not compatible, so it's not safe to build against both the 2.3 libraries as well as use the 2.6 compiler. This breaks (for example) external/webrtc, which would be fixed by https://android-review.googlesource.com/#/c/107751, but now I'm concerned that there might be lots of other places affected by this.

Tomorrow I'll take a look and see how many places this affects.

Patch-set: 5
Label: Verified=0