Update patch set 2
Patch Set 2: Code-Review-2

Not ready for submit yet, but I was looking for an initial set of thoughts here. Do you think this is reasonable? Any concerns?

AFAIK the major missing pieces are:

1. Supporting all flag combinations (this only works if you have accessors on and is missing some pieces otherwise)

2. Gating this by some kind of flag (generate_intdef_annotations or similar).

Let me know if you think this is a good idea - I think it could definitely be useful.

Patch-set: 2
Label: Code-Review=-2