Update patch set 4
Patch Set 4:

I think it's sensible to default the IntDef generation to off, and then ask the user to manually include the required library if they turn it on.

Also, it seems if we're using shell class, then you can directly turn it into an annotation, and client code will still be able to refer to the int constants in there.

Patch-set: 4
Label: Verified=0