Update patch set 5
Patch Set 5: -Code-Review

OK: I think this is feature-complete. Please take a full look. Note that I am pretty inexperienced with both C++ and this code base so a close look is encouraged :)

The feature is now gated by a default-off flag. It should also work with both C and Java-style enums. For C-style enums, we declare the annotation using the same name as the outer interface that we would have otherwise used for Java-style enums. (Is this not safe to do?) For Java-style enums, we repurpose the outer interface as the annotation, as Max wisely suggested.

Patch-set: 5