Generate a CREATOR for each Parcelable message.
This is definitely less ideal from a class/method count perspective,
because it requires a new inner class with two methods for each
message, and because most apps have proguard rules that will ensure
that CREATOR classes are retained.

However, it is required to be able to use nano protos inside of AIDL
files, as the autogenerated AIDL code fails to compile otherwise. This
is a substantial benefit as it allows for backwards-compatible
parameters and return types in AIDL methods along the lines of
safeparcel.

Bug: 19084705
Change-Id: I66a2c0424b96cf8ff6b631b186cc4f9407dfc1f4