Update patch set 1
Patch Set 1:

the problem is that the null pointer exceptions don't happen at the point you set things to null, they happen later when you invoke other methods on the protos. This violates the "fail fast" good practice for apis. Because nano is field-based, there's no way to do any kind of enforcement at the time fields are set.

Patch-set: 1