Update patch set 3
Patch Set 3:

=== Finished presubmit run: L05900000108018841 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/499050?revision=3




NOTE: Since this build started, a new topic was set or a new patch was uploaded.

TreeHugger finished with: 13 passed, 2 failed.

FAIL: Failed to build aosp-build-tools/linux/4372208 (LKGB). See https://android-build.googleplex.com/builds/pending/P4837073/linux/latest
FAIL: Failed to build aosp-build-tools/linux/4372562 (ToT). See https://android-build.googleplex.com/builds/pending/P4837074/linux/latest. WARNING: Your change was built on top of build 4372562, which is likely to fail because the previous build is broken. Please check the continuous build at http://go/abbb/aosp-build-tools/grid and retry your patch only when linux is green.
WARNING: The tip-of-tree(4372562) for aosp-build-tools/linux may be broken.
Presubmit builds based on the tip-of-tree (4372562) and the
last-known-good-build(4372208) also fail. This could be due to the
pre-existing error, a new error in this change, or both.


WARNING: The tip-of-tree(4372561) for aosp-master/aosp_arm64-eng may be broken.
The presubmit build passes with the last-known-good-build(4372162), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4372561) for aosp-master/aosp_x86-eng may be broken.
The presubmit build passes with the last-known-good-build(4372213), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



















Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 3
Tag: autogenerated:TreeHugger:Verify-Stale-Fail
Reviewer: Treehugger Robot <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=0