protobuf: add lite ndk libc++ static target
stlport has been deprecated and will be removed at some point
(https://developer.android.com/ndk/guides/cpp-support.html). Moreover,
it does not have support for C++11 and forces users of the target to
also not use C++11.

Introduce a target for ndk libc++ which is modern and the future of
the STL NDK.

Test: Compile works - checked that new target can be used
Change-Id: Ie35d01787b64056e9f61b6fc31902deba67d6717