Build libprotobuf-java-lite only using core Java SDK APIs
libprotobuf-java-lite is now built against core Java APIs which excludes
Android-only APIs (classes under android.* packages) and private APIs.
"core_current" is a pseudo SDK which is a core Java subset of Android
SDK.

However, this lib has been using sun.misc.Unsafe which is a private API
and dependency to the class doesn't seem to be able to be cut easily in
the near future. In order to make it possible to build the lib with the
SDK only, a stub class of Unsafe with required methods only has been
added to the source tree. Note that methods in this stub class is never
invoked at runtime; it only exist to provide symbols during the
build-time.

At runtime, if Unsafe is provided by the OS, then the real class is
loaded. Even when the OS does not provide Unsafe, the stub class is not
used because protobuf relies on the existence of a static field of type
Unsafe to reflectively check if the class is really available or not.
The stub class does not have the static field.

Bug: 73204840
Test: mma -j under external/protobuf

Change-Id: If55130c0b8b519ab355527b27429faac76e7494f