Build libprotobuf-java-lite only using core Java SDK APIs
libprotobuf-java-lite is now built against core Java APIs which excludes
Android-only APIs (classes under android.* packages) and private APIs.
"core_current" is a pseudo SDK which is a core Java subset of Android
SDK.

However, this lib has been using sun.misc.Unsafe which is a hidden API
and dependency to the class doesn't seem to be able to be cut easily in
the near future. In order to make it possible to build the lib with the
SDK only, a stubs library called libcore_private.stubs is defined to
provide build-time only information on the hidden APIs exclusively to
protobuf.

At runtime, if Unsafe is provided by the OS, then the real class is
loaded. Even when the OS does not provide Unsafe, the stub class is not
used because protobuf relies on the existence of a static field of type
Unsafe to reflectively check if the class is really available or not.
The stub class does not have the static field.

Bug: 73204840
Test: m -j libprotobuf-java-lite is successful and does not show any
link-type check warning

Change-Id: I2fce2662cc65d4e57fc75bbf30ebf9f7d47c6775