Build libprotobuf-java-lite only using core Java SDK APIs
libprotobuf-java-lite is now built against core Java APIs which excludes
Android-only APIs (classes under android.* packages) and private APIs.
"core_current" is a pseudo SDK which is a core Java subset of Android
SDK.

However, this lib has been using sun.misc.Unsafe which is a hidden API
and dependency to the class doesn't seem to be able to be cut easily in
the near future. In order to make it possible to build the lib with the
SDK only, a stubs library called libcore_private.stubs is defined to
provide build-time only information on the hidden APIs exclusively to
protobuf.

At runtime, the Unsafe class is provide by the OS and protobuf is linked
to the class loaded there.

Bug: 73204840
Test: m -j libprotobuf-java-lite is successful and does not show any
link-type check warning

Change-Id: I2fce2662cc65d4e57fc75bbf30ebf9f7d47c6775