Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> Okay as a short to medium term workaround, but in the medium to long term, libcore_private.stubs should move away from adding additional classes in libcore packages. If the stubs can't be built with the latest language level, that might prevent us adding new APIs whose signatures depend on the latest language level.

I too am not fully satisfied with this CL. In fact, I am trying to make this stub a target-only module. The error happens while building for host, but this stub isn't actually needed in the host build since JDK already provide Unsafe; it is built just because of the dependency from libprotobuf-java-lite.

Patch-set: 2