Update patch set 2
Patch Set 2:

> I too am not fully satisfied with this CL. In fact, I am trying to make this stub a target-only module. The error happens while building for host, but this stub isn't actually needed in the host build since JDK already provide Unsafe; it is built just because of the dependency from libprotobuf-java-lite.

Perhaps libprotobuf-java-lite needs to have a javac flag of --patch-module=java.base=<path to stubs jar>; that way it should be able to compile with OpenJDK 9 javac -target 1.9.

In the longer term of course, we should talk to the proto folks about replacing the dependency on Unsafe with a supported API that fulfills their needs.

Patch-set: 2