Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Does this mean two copies of libprotobuf can be loaded in the same process? 
> 
> Yes, two different variants/versions of libprotobuf can be loaded in the same vendor process which depends on both libmediandk(LLNDK) and libprotobuf-cpp-lite(VNDK).
> 
> > If so, we've had ODR problems with that repeatedly in the past, for example b/36272398 and b/36900282.
> 
> This CL only reflects the current status.

In my understanding, this is different from ODR problem. The two libraries are loaded into the distinct linker namespaces, so symbol A in a namespace ns1 does not collide with the same symbol A in another namespace ns2.

Double loadable can be a problem if the library is written with an assumption that it is the only entity in the process that can hold a process-global resource such as mutex, file descriptor, etc.

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>