Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I'd be in favor of keeping the C++ stuff out of the generic plugin library, lest it start to grow more dependencies.  But I don't have a strong opinion either way.

I'm fine with keeping it separate, but then it needs a name that makes it clear what the difference is from libprotoc.  If it's just for kythe, maybe libprotoc-kythe, and use visibility to restrict it to be used only by external/kythe?

Patch-set: 2
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>