Update patch set 1
Patch Set 1: -Code-Review

Colin, thanks.
Jiyong, I thought this was necessary for my change to linkerconfig and that's why I changed the topic for this change to match with mine. As Colin pointed out, I switched to libprotobuf-cpp-lite for linkerconfig instead of libprotobuf-cpp-lite_static and it works. I'm not sure this is still necessary for your change to init.

Patch-set: 1
Reviewer: Gerrit User 1415877 <1415877@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0