Update patch set 2
Patch Set 2:

> why can a vendor APK access vendor/lib but not the VNDK?

A vendor APK can access /vendor/lib and /system/lib/vndk-sp, but not /system/lib/vndk. This is because VNDK libs (except for VNDK-SP libs) are by default not available for double loading in a process. 

> Putting them in the VNDK was supposed to solve that.

It is already in VNDK, but not in VNDK-SP. We could satisfy their need by promoting this library to VNDK-SP (thus allowing double loading), but we have been trying to keep VNDK-SP libs as small as possible since double loading is just scary.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0