Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Device code shouldn't be using full proto. They should be using lite. The full proto library has a LOT of bloat.

Hmm, ok. I think I agree with it. But the reality is that libprotobuf-cpp-full has already been available (and used) for device for a long time in AOSP. Making libprotobuf-cpp-full a host-only library seems to be a good goal, but IMO that is out of scope of this change.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0